requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data versioning
dvc>=3.0.0

# Experiment tracking
mlflow>=2.0.0

# API framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0

# Data processing
opencv-python>=4.8.0
Pillow>=9.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Docker and deployment
docker>=6.0.0

# Utilities
tqdm>=4.65.0
click>=8.0.0 